backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
opencv-python
pillow